Share request credential parsing between REST and MDS.
Extract RequestCredentials so both ApplicationController and MDS auth use
the same Authorization header split, User construction, and JWT blacklist check.
